+++ src/js/libs/react-dnd.js
@@ -1,0 +1,10 @@
+export const ItemTypes = {
+  TAB: 'tab'
+}
+
+export const collect = (connect, monitor) => ({
+  connectDropTarget: connect.dropTarget(),
+  canDrop: monitor.canDrop(),
+  isDragging: !!monitor.getItem(),
+  isOver: monitor.isOver({ shallow: true })
+})
